Clear events.email instead of deleting the whole event row
Onetime::DeleteEventsWithEmail removed residual buyer PII from the legacy
events.email column by deleting every events row where email was present.
But per the investigation in #5383, those rows are not dead tracking cruft:
each one is a recurring-subscription-charge `purchase` event. They are
created on every recurring charge by Subscription#create_purchase_event,
which dups the subscription's original purchase event and (until #5383)
copied an old buyer email forward.

Those purchase events are live, referenced records:

- Purchase::ChargeEventsHandler, Purchase#refundable, and Charge::Disputable
  each look up Event.where(purchase_id:, event_name: "purchase").last and dup
  it to build the refund/chargeback/dispute/service-charge event. Deleting the
  purchase event makes these silently no-op (guarded by `if purchase_event`),
  so those events stop being recorded for the affected charges.
- Subscription#create_purchase_event dups the original/template purchase event
  for each new charge.
- Sales/revenue analytics and reports count `purchase` events.

So delete_all here is data loss, not PII cleanup. The only reader of the
email column is GdprBuyerErasureService#anonymize_events!, which nulls the
column rather than deleting rows. This change does the same: null email in
batches, keep the row. That removes the residual PII and still unblocks
dropping the column later, without destroying referenced purchase events.

Renames the task to Onetime::ClearEventsEmail to match the new behavior, and
updates the spec to assert the row survives with email nil (so it fails if
reverted to delete_all).
