+++ spec/services/onetime/clear_events_email_spec.rb
@@ -2,24 +2,26 @@
 
 require "spec_helper"
 
-describe Onetime::DeleteEventsWithEmail do
+describe Onetime::ClearEventsEmail do
   describe ".process" do
-    it "deletes events that have an email and keeps events without one" do
+    it "nulls the email on events that have one and keeps the rows" do
       with_email = create(:event, email: "buyer@example.com")
       without_email = create(:event, email: nil)
 
       described_class.process
 
-      expect(Event.exists?(with_email.id)).to be(false)
+      expect(Event.exists?(with_email.id)).to be(true)
+      expect(with_email.reload.email).to be_nil
       expect(Event.exists?(without_email.id)).to be(true)
     end
 
-    it "deletes across multiple batches and returns the total number deleted" do
-      create_list(:event, 3, email: "buyer@example.com")
+    it "clears across multiple batches and returns the total number cleared" do
+      cleared = create_list(:event, 3, email: "buyer@example.com")
       create(:event, email: nil)
 
       expect(described_class.process(batch_size: 1)).to eq(3)
       expect(Event.where.not(email: nil).count).to eq(0)
+      cleared.each { expect(_1.reload.email).to be_nil }
     end
 
     it "pauses for replica lag, allowing up to 10 seconds of delay, between batches" do
